src/requirements.txt
flask==2.0.2
pydantic==1.8.2
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
requests==2.26.0
multidict==5.2.0
pydantic-sqlalchemy==0.0.9
flask-pydantic[email]==0.8.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
flask-injector==0.13.0
flask-jwt-extended==4.3.1
pyyaml==6.0
redis==3.5.3
Werkzeug==2.0.2
injector==0.18.4
marshmallow==3.14.0
click==8.0.3
gevent==21.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
apispec[marshmallow]==5.1.1
apispec-webframeworks==0.5.2
flask-swagger-ui==3.36.0
